requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==4.2.9
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
idna==3.6
numpy==1.26.3
pandas==2.1.4
pycparser==2.21
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-memcached==1.62
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.4